# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TiendaPOS API",
        "description": "SaaS POS Multi-País para LATAM. Soporta 9 países (VE, CO, MX, EC, AR, PE, CL, BO, UY) con reglas fiscales, IVA, monedas y métodos de pago locales.",
        "contact": {
            "email": "soporte@tiendapos.com"
        },
        "version": "2.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/tienda": {
            "get": {
                "tags": [
                    "Configuración"
                ],
                "summary": "Obtener configuración de la tienda",
                "description": "Devuelve la configuración de la tienda.",
                "operationId": "4d15962a8a6ac61011bd0d3d6909cacf",
                "responses": {
                    "200": {
                        "description": "Configuración de la tienda"
                    },
                    "404": {
                        "description": "Tienda no configurada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Configuración"
                ],
                "summary": "Actualizar configuración de la tienda",
                "description": "Actualiza la configuración de la tienda (solo admin).",
                "operationId": "9c3b1b55f2da91ab8d41d340a6b6a6a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "razon_social": {
                                        "type": "string"
                                    },
                                    "nombre_comercial": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "prefijo_factura": {
                                        "type": "string"
                                    },
                                    "decimales_precio": {
                                        "type": "integer"
                                    },
                                    "es_agente_igtf": {
                                        "type": "boolean"
                                    },
                                    "alicuota_igtf": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tienda actualizada"
                    },
                    "404": {
                        "description": "Tienda no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum token de autenticación. Obténlo via POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}